Only clear exception, if parsed node is longer than node that failed
In catch_exceptions, the exception only gets cleared, if the node that
was parsed in the alternative was longer than the node that failed.
This allows for nodes to succeed, without clearing the exception, while
it also makes sure that the exception does not get cleared every time,
thus losing the error information.