[WIP]: Have a stack of errors
`throw` appends the error to the stack and `catch` currently pops
all the errors in from stack, until an an exception type is found
that is getting caught.

Note to self:
I need to rethink, when exception are cleared. Clearing it every
time does not really work and clearing it when the parsed sentence
is longer than the error also does not work. See commit history
for examples.